Migrated to decimal

+++ Math/Nodes/Functions/Binary/SubstractionNode.cs
@@ -11,7 +11,7 @@ namespace Math.Nodes.Functions.Binary
         public INode RightNode { get; set; }
         public INode LeftNode { get; set; }
 
-        public double Evaluate()
+        public decimal Evaluate()
         {
             return LeftNode.Evaluate() - RightNode.Evaluate();
         }
