+++ Math/Nodes/Functions/Binary/SumNode.cs
@@ -5,7 +5,7 @@ namespace Math.Nodes.Functions.Binary
         public INode RightNode { get; set; }
         public INode LeftNode { get; set; }
         
-        public virtual double Evaluate()
+        public virtual decimal Evaluate()
         {
             return LeftNode.Evaluate() + RightNode.Evaluate();
         }
